requirements.txt
groq
uvicorn
fastapi
streamlit
python-dotenv
pydantic
langchain
langchain-community
pypdf2
pypdf
pinecone-client
python-dotenv
langchain-groq
langchain-openai
sentence_transformers
protoc_gen_openapiv2
langchain-pinecone
faiss-cpu
anthropic
vertexai
llama-index-core
llama-index-readers-file
llama-index-embeddings-huggingface==0.2.3
llama-index-llms-groq
llama-index-llms-vertex
llama-index-llms-anthropic